Split "plug" function into a "plug" and an "attach" function
Recent versions of docker expect two functions, one named "plug" and the
other named "attach". This commit changes the code so that both functions
exist and do the right thing to make Docker happy. Without the patch, the
correct association of interfaces and network namespaces is broken (i.e.
the driver will try to add an interface to a namespace that does nut exist
yet).