Don't make the driver crash nova-compute if veth if already exists
In its current form, the opencontrail driver for docker will try to create
a veth interface even if it's already there. This will result in an error
that will make nova-compute crash. In fact, nova-compute is unrestartable
with the current driver. This patch adds a check to see if the veth if is
already there and does not try to add it a second time if it already is.